fixed cursor blinking, must be 64-bit counter now
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@4853 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/console/OSDConsoleRenderer.hh
@@ -44,7 +44,7 @@ protected:
 	  * that image's alpha channel is used instead.
 	  */
 	static const int CONSOLE_ALPHA = 180;
-	static const int BLINK_RATE = 500000; // us
+	static const unsigned long long BLINK_RATE = 500000; // us
 	static const int CHAR_BORDER = 4;
 
 	enum Placement {
@@ -60,7 +60,7 @@ protected:
 	std::auto_ptr<FilenameSetting> fontSetting;
 	std::auto_ptr<Font> font;
 	bool blink;
-	unsigned lastBlinkTime;
+	unsigned long long lastBlinkTime;
 	unsigned lastCursorX;
 	unsigned lastCursorY;
 	unsigned destX;
